just a bit of cleanup
