+++ .rspec_status
@@ -1,9 +1,9 @@
 example_id                                           | status | run_time        |
 ---------------------------------------------------- | ------ | --------------- |
-./spec/lib/gp_webpay/config_spec.rb[1:1:1]           | passed | 0.00192 seconds |
-./spec/lib/gp_webpay/config_spec.rb[1:1:2]           | passed | 0.00056 seconds |
-./spec/lib/gp_webpay/config_spec.rb[1:2:1]           | passed | 0.00703 seconds |
-./spec/lib/gp_webpay/config_spec.rb[1:2:2]           | passed | 0.00019 seconds |
-./spec/lib/gp_webpay/payment_attributes_spec.rb[1:1] | passed | 0.00022 seconds |
-./spec/lib/gp_webpay/payment_attributes_spec.rb[1:2] | passed | 0.00046 seconds |
-./spec/lib/gp_webpay/payment_spec.rb[1:1]            | passed | 0.00097 seconds |
+./spec/lib/gp_webpay/config_spec.rb[1:1:1]           | passed | 0.00154 seconds |
+./spec/lib/gp_webpay/config_spec.rb[1:1:2]           | passed | 0.00055 seconds |
+./spec/lib/gp_webpay/config_spec.rb[1:2:1]           | passed | 0.00352 seconds |
+./spec/lib/gp_webpay/config_spec.rb[1:2:2]           | passed | 0.00052 seconds |
+./spec/lib/gp_webpay/payment_attributes_spec.rb[1:1] | passed | 0.00028 seconds |
+./spec/lib/gp_webpay/payment_attributes_spec.rb[1:2] | passed | 0.00058 seconds |
+./spec/lib/gp_webpay/payment_spec.rb[1:1]            | passed | 0.00106 seconds |
