+++ bin/console
@@ -3,5 +3,5 @@
 require "bundler/setup"
 require "gp_webpay"
 
-require "irb"
-IRB.start
+require "pry"
+Pry.start
