+++ spec/lib/gp_webpay/payment_attributes_spec.rb
@@ -1,27 +1,3 @@
-class Payment
-  include GpWebpay::Payment
-
-  def initialize
-    self.redirect_url = "https://test"
-  end
-
-  def order_number
-    1001
-  end
-
-  def payment_type
-    "default"
-  end
-
-  def amount_in_cents
-    100
-  end
-
-  def currency
-    103
-  end
-end
-
 RSpec.describe GpWebpay::PaymentAttributes do
   let(:payment) { Payment.new }
   let(:payment_attributes) { GpWebpay::PaymentAttributes.new(payment) }
